+++ css/form.css
@@ -1,0 +1,364 @@
+.about_box2 {
+  box-sizing: border-box;
+}
+.about_box2 img {
+  width: 100%;
+  height: auto;
+  display: block;
+  object-fit: cover;
+  border-radius: 10px;
+}
+/* Tabs */
+.tab {
+  overflow: hidden;
+  border: 1px solid #ccc;
+  background-color: #f1f1f1;
+}
+
+.tab button {
+  background-color: inherit;
+  float: left;
+  border: none;
+  outline: none;
+  cursor: pointer;
+  padding: 14px 16px;
+  transition: 0.3s;
+}
+
+.tab button:hover {
+  background-color: #ddd;
+}
+
+.tab button.active {
+  background-color: #ccc;
+}
+
+.tabcontent {
+  display: none;
+  justify-content: center;
+}
+
+#ASA {
+  text-align: center;
+}
+
+/* Filter Bar */
+.filter-bar {
+  width: 60%;
+  margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  position: relative;
+}
+
+#filter-club {
+  width: 100%;
+  margin: 0;
+  padding: 8px;
+  justify-self: left;
+}
+
+.club-input-wrapper {
+  position: relative;
+  width: 100%;
+  box-sizing: border-box;
+}
+
+.autocomplete-items,
+.autocomplete-club {
+  position: absolute;
+  left: 0;
+  width: 100%;
+  background: white;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  z-index: 1000;
+}
+
+.autocomplete-items:empty {
+  box-shadow: none;
+  display: none;
+}
+
+.autocomplete-items,
+.autocomplete-club div {
+  padding: 8px;
+  cursor: pointer;
+  transition: 0.2s;
+  text-align: left;
+}
+
+.autocomplete-items div:hover,
+.autocomplete-items div.active,
+.autocomplete-club div:hover,
+.autocomplete-club div.active {
+  background-color: #f0f0f0;
+}
+
+.filter-row {
+  display: flex;
+  gap: 10px;
+  flex-wrap: wrap;
+}
+
+.filter-row select,
+.filter-row input,
+.filter-row button {
+  flex: 1;
+  padding: 8px;
+  min-width: 100px;
+  box-sizing: border-box;
+}
+/*Input Section*/
+.input_box1 {
+  box-sizing: border-box;
+}
+
+.input_box1 img {
+  width: 100%;
+  height: auto;
+  display: block;
+  object-fit: cover;
+  border-radius: 10px;
+}
+
+.input_box2 p {
+  margin: 5px;
+}
+
+/* Form container */
+.form-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.swimmer-container {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  align-items: flex-start;
+}
+
+.individual-input {
+  position: relative;
+  padding: 20px;
+  width: 350px;
+  margin: 20px auto;
+  border: 1px solid #ccc;
+  border-radius: 8px;
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
+  margin-bottom: 15px;
+  font-size: 1.2em;
+  color: #333;
+}
+
+.individual-input.collapsed {
+  flex: 0 0 auto;
+}
+
+.form-input {
+  margin-bottom: 15px;
+}
+
+.form-input label {
+  display: block;
+  font-weight: bold;
+  margin-bottom: 5px;
+}
+
+.form-input input,
+.form-input select {
+  width: 100%;
+  padding: 8px;
+  font-size: 1em;
+  border: 1px solid #bbb;
+  border-radius: 4px;
+}
+
+.swimmer-container_sc {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.swimmer-input {
+  width: 60%;
+  background-color: var(--light-grey);
+  border: none;
+  border-radius: 30px;
+  padding: 0;
+  display: flex;
+  align-items: center;
+  margin-bottom: 10px;
+  overflow: hidden;
+}
+
+label {
+  font-size: small;
+  font-weight: bold;
+  flex: 1;
+}
+
+.divider {
+  height: 30px;
+  width: 2px;
+  background-color: #000;
+  margin: 0 10;
+}
+
+input[type="text"] {
+  flex: 2;
+  margin: 3px 20px 3px 10px;
+  border: none;
+  background-color: var(--light-grey);
+}
+
+input[type="text"]:focus {
+  outline: none;
+}
+
+#addSwimmerButton {
+  align-self: flex-end;
+  margin-right: 10%;
+}
+
+#addSwimmerButton:hover {
+  opacity: 70%;
+}
+
+.delete-button-container {
+  display: flex;
+  align-items: center;
+  margin: 0 20px 10px 0;
+}
+
+.fa-trash {
+  color: red;
+  position: absolute;
+  top: 0;
+  right: 0;
+  padding: 10px 10px 0 0;
+  transition: 0.3s;
+  cursor: pointer;
+}
+
+.fa-trash:hover {
+  color: darkred;
+  transition: 0.5s;
+}
+
+.edit-button {
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  padding: 10px;
+  color: #000;
+  border: none;
+  cursor: pointer;
+  transition: 0.3s ease;
+}
+
+.edit-button:hover {
+  color: grey;
+  transition: 0.5s;
+}
+
+.btn-delete {
+  font-size: 2em;
+  font-weight: bold height 30px;
+  line-height: 30px;
+  width: 30px;
+  background-color: red;
+  color: #000;
+  border-radius: 50%;
+  border: none;
+  cursor: pointer;
+}
+
+.btn-delete:hover {
+  background-color: darkred;
+}
+wrapper {
+  width: 100%;
+  position: relative;
+  display: flex;
+  justify-content: center;
+}
+.new-one {
+  position: relative;
+}
+
+#addSwimmerButton {
+  align-self: flex-end;
+  margin-right: 10%;
+}
+/* Radio Buttons */
+input[type="radio"] {
+  border: 0 !important;
+  clip: rect(1px, 1px, 1px, 1px);
+  height: 1px !important;
+  overflow: hidden;
+  padding: 0 !important;
+  position: absolute !important;
+  width: 1px !important;
+}
+
+input[type="radio"]:checked {
+  background-color: blue;
+  border: 6px solid white;
+  box-shadow: 0 0 0 2px blue;
+}
+
+.radio-label {
+  color: black;
+  font-size: 1em;
+  opacity: 70%;
+  display: inline-flex;
+  align-items: center;
+  margin-right: 20px;
+}
+
+.radio-label::before {
+  content: "";
+  display: block;
+  height: 1.25em;
+  width: 1.25em;
+  border: solid 1px #ccc;
+  border-radius: 50%;
+  margin-right: 0.5em;
+  margin-left: 20px;
+  transition: border ease-in 150ms, box-shadow ease-in 150ms;
+}
+
+input[type="radio"]:checked + label {
+  color: var(--dark-blue);
+  opacity: 100%;
+}
+
+input[type="radio"]:hover + label {
+  opacity: 100%;
+}
+
+input[type="radio"]:checked + label::before {
+  background: radial-gradient(
+    0.75em circle at center,
+    var(--dark-blue) 50%,
+    transparent 60%
+  );
+  box-shadow: 0 0 1em 0 rgba(10, 150, 255, 0.75);
+  border-color: currentColor;
+}
+
+input[type="radio"]:hover + label::before {
+  box-shadow: 0 0 1em 0 rgba(10, 150, 255, 0.75);
+  border-color: currentColor;
+}
+.buttons-container > * {
+  margin-left: 5px;
+}
+.hidden {
+  display: none;
+}
