+++ css/variables.css
@@ -1,0 +1,8 @@
+:root {
+  --white: #fff;
+  --light-blue: #2cd8ff;
+  --dark-blue: #000072;
+  --black: #000;
+  --light-grey: #dadada;
+  --dark-grey: #b1b2b5;
+}
