+++ js/table.js
@@ -1,0 +1,48 @@
+export function tbl_create(swimmer_list,index) {
+  const tbl_section = document.getElementById("results");
+  const tbl = document.createElement("table");
+  tbl.className = "results-table";
+
+  /*Make the Header*/
+  const tbl_header = document.createElement("thead");
+  ["Stroke", "Name", "Time"].forEach((text) => {
+    const head = document.createElement("th");
+    head.textContent = text;
+    tbl_header.appendChild(head);
+  });
+  tbl.appendChild(tbl_header);
+
+  /*Add table contents*/
+  for (let n = 0; n < swimmer_list.length - 1; n++) {
+    const tr = document.createElement("tr");
+    swimmer_list[n].forEach((text) => {
+      const td = document.createElement("td");
+      td.textContent = text;
+      tr.appendChild(td);
+    });
+    tbl.appendChild(tr);
+  }
+  const tr = document.createElement("tr");
+  let last_row = swimmer_list[swimmer_list.length - 1];
+  const td1 = document.createElement("td");
+  td1.colSpan = "2";
+  td1.textContent = last_row[0];
+  const td2 = document.createElement("td");
+  td2.textContent = last_row[1];
+  tr.appendChild(td1);
+  tr.appendChild(td2);
+  tbl.appendChild(tr);
+
+  let tbl_wrapper = document.createElement("div");
+  tbl_wrapper.className = "table-container";
+  tbl_wrapper.appendChild(tbl);
+  tbl_section.appendChild(tbl_wrapper);
+  if (index === 0) {
+    document.getElementById("table-row-top").appendChild(tbl_wrapper);
+  } else if (index === 5) {
+    document.getElementById("table-row-bottom").appendChild(tbl_wrapper);
+  } else {
+    document.getElementById("table-row-mid").appendChild(tbl_wrapper);
+  }
+  tbl_wrapper.scrollIntoView({ behavior: "smooth" });
+}
